requirements.txt
Flask~=3.1.0
Werkzeug~=3.1.3
pillow~=11.2.1
gunicorn~=21.2.0
